+++ src/app/data/db.service.ts
@@ -154,6 +154,11 @@ export class DbService {
     return await call(this.worker, DataMethods.GetGeoReferences);
   }
 
+  // True if the dataset has geo points. False if we avoid geolocation
+  public async hasGeoPoints(): Promise<boolean> {
+    return await call(this.worker, DataMethods.HasGeoPoints);
+  }
+
   public async gpsToPoint(gpsCoord: GpsCoord): Promise<Point> {
     return await call(this.worker, DataMethods.GpsToPoint, gpsCoord);
   }